+++ data-parts.go
@@ -37,9 +37,9 @@ func loadPartsTable(writer http.ResponseWriter, workplaceIds string, dateFrom ti
 	data.DataTableSearchTitle = getLocale(email, "data-table-search-title")
 	data.DataTableInfoTitle = getLocale(email, "data-table-info-title")
 	data.DataTableRowsCountTitle = getLocale(email, "data-table-rows-count-title")
-	companyNameSync.Lock()
-	loc, err := time.LoadLocation(location)
-	companyNameSync.Unlock()
+	locationSync.RLock()
+	loc, err := time.LoadLocation(cachedLocation)
+	locationSync.RUnlock()
 	addPartTableHeaders(email, &data)
 	for _, record := range packageRecords {
 		addPartTableRow(record, &data, db, loc)