+++ data-system-stats.go
@@ -33,9 +33,9 @@ func loadSystemStatsTable(writer http.ResponseWriter, dateFrom time.Time, dateTo
 	data.DataTableSearchTitle = getLocale(email, "data-table-search-title")
 	data.DataTableInfoTitle = getLocale(email, "data-table-info-title")
 	data.DataTableRowsCountTitle = getLocale(email, "data-table-rows-count-title")
-	companyNameSync.Lock()
-	loc, err := time.LoadLocation(location)
-	companyNameSync.Unlock()
+	locationSync.RLock()
+	loc, err := time.LoadLocation(cachedLocation)
+	locationSync.RUnlock()
 	addSystemTableHeaders(email, &data)
 	for _, record := range systemRecords {
 		addSystemTableRow(record, &data, loc)