+++ data-users.go
@@ -36,9 +36,9 @@ func loadUsersTable(writer http.ResponseWriter, workplaceIds string, dateFrom ti
 	data.DataTableSearchTitle = getLocale(email, "data-table-search-title")
 	data.DataTableInfoTitle = getLocale(email, "data-table-info-title")
 	data.DataTableRowsCountTitle = getLocale(email, "data-table-rows-count-title")
-	companyNameSync.Lock()
-	loc, err := time.LoadLocation(location)
-	companyNameSync.Unlock()
+	locationSync.RLock()
+	loc, err := time.LoadLocation(cachedLocation)
+	locationSync.RUnlock()
 	addUserTableHeaders(email, &data)
 	for _, record := range userRecords {
 		addUserTableRow(record, &data, db, loc)